pyproject.toml
[project]
name = "exponent_server_sdk_async"
version = "2.1.6"
description = "Expo Server SDK for Python async"
authors = [
    {name = "Bahadır Araz", email = "bahadiraraz@protonmail.com"},
    {name = "Expo Team", email = "exponent.team@gmail.com"},
]
keywords = ["expo", "sdk", "async", "push notifications"]

dependencies = [
    "six",
    "httpx>=0.27.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/Learnity-AI/expo-server-sdk-python-async"
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
